Animations and Launch Angles tweak
Switched method of getting launch angle. This one is *infinitely* better since it awards the player for hitting the ball down the middle (when under it at least) by creating a homerun-like angle, compared to last time where you had to be under the ball by a huge distance to hit a homerun, which isn't fair for players who hit it down the middle (again, when they're under it at least).